refactor: reorganise internal logic flow
