+++ .changeset/purple-cooks-compete.md
@@ -1,0 +1,5 @@
+---
+"openapi-ts-json-schema": minor
+---
+
+Returned `metaData.schema` property replaced by `metaData.originalSchema` holding the original dereferenced JSON schema
