+++ src/utils/makeTsJsonSchema/patchJsonSchema.ts
@@ -1,6 +1,9 @@
 import traverse from 'json-schema-traverse';
-import type { JSONSchema, SchemaPatcher } from '../types';
+import type { JSONSchema, SchemaPatcher } from '../../types';
 
+/**
+ * Patch generated schemas with a user-provided patch function
+ */
 export function patchJsonSchema(
   schema: JSONSchema,
   schemaPatcher?: SchemaPatcher,
