+++ src/utils/makeTsJsonSchemaFiles.ts
@@ -1,5 +1,5 @@
 import { makeTsJsonSchema, saveFile } from '.';
-import type { SchemaMetaDataMap } from '../types';
+import type { SchemaMetaDataMap, SchemaPatcher } from '../types';
 
 /**
  * Save TS JSON schema files with the expected naming conventions
@@ -7,15 +7,18 @@ import type { SchemaMetaDataMap } from '../types';
 export async function makeTsJsonSchemaFiles({
   schemaMetaDataMap,
   refHandling,
+  schemaPatcher,
 }: {
   schemaMetaDataMap: SchemaMetaDataMap;
   refHandling: 'inline' | 'import' | 'keep';
+  schemaPatcher?: SchemaPatcher;
 }) {
   for (const [_, metaData] of schemaMetaDataMap) {
     const tsSchema = await makeTsJsonSchema({
       metaData,
       schemaMetaDataMap,
       refHandling,
+      schemaPatcher,
     });
 
     const { schemaAbsolutePath } = metaData;
